WebApp-Docker/web/requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
Flask==3.1.0
Flask-Session==0.5.0
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.32.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==80.4.0
urllib3==2.4.0
Werkzeug==3.1.3